Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.3
scikit-learn==1.7.2
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
